# optimization: reduced memory allocations for append().