As the size of `lines` is known before running append(), make() reserves the capacity for `records` as same as it.

## Benchmark

There are simple tests and bechmarks for recordset-creation in https://github.com/fujiwara/fluent-agent-hydra/commit/74f186ea4dcf2e5aff867b4eda17771c6c33bc02.

### Before

```console
$ go test -run=nonthingplease -bench NewFluentRecordSetJSON -count 5 -benchmem -benchtime=3s
PASS
BenchmarkNewFluentRecordSetJSON-8          50000            101019 ns/op           19424 B/op        627 allocs/op
BenchmarkNewFluentRecordSetJSON-8          50000            101360 ns/op           19424 B/op        627 allocs/op
BenchmarkNewFluentRecordSetJSON-8          50000            100986 ns/op           19424 B/op        627 allocs/op
BenchmarkNewFluentRecordSetJSON-8          50000            101244 ns/op           19424 B/op        627 allocs/op
BenchmarkNewFluentRecordSetJSON-8          50000            100513 ns/op           19424 B/op        627 allocs/op
ok      github.com/fujiwara/fluent-agent-hydra/hydra    30.368s
```

### After

```console
$ go test -run=nonthingplease -bench NewFluentRecordSetJSON -count 5 -benchmem -benchtime=3s
PASS
BenchmarkNewFluentRecordSetJSON-8          50000             99576 ns/op           19088 B/op        623 allocs/op
BenchmarkNewFluentRecordSetJSON-8          50000            100329 ns/op           19088 B/op        623 allocs/op
BenchmarkNewFluentRecordSetJSON-8          50000             99746 ns/op           19088 B/op        623 allocs/op
BenchmarkNewFluentRecordSetJSON-8          50000             99744 ns/op           19088 B/op        623 allocs/op
BenchmarkNewFluentRecordSetJSON-8          50000             99933 ns/op           19088 B/op        623 allocs/op
ok      github.com/fujiwara/fluent-agent-hydra/hydra    30.064s
```